refactor: group proviers
